requirements.txt
PyQt6>=6.6.0
PyQt6-Charts>=6.5.0
PyQt6-Qt6>=6.6.0
PyQt6-sip>=13.6.0
qrcode[pil]>=7.4.2
opencv-python>=4.8.1
pillow>=10.1.0
numpy>=1.26.2
pandas>=2.1.3
matplotlib>=3.8.2
SQLAlchemy>=2.0.23